dts: bindings: xen: add xen,xen.yaml file
Add yaml file for 'xen,xen', because without it an appropriate
'CONFIG_DT_HAS_XEN_XEN_ENABLED' isn't generated.

Include a new generic xen,xen.yaml file to xen,xen-4.15.yaml,
because the first has properties for both.

Add subdirectory of 'xen' when we have Xen Hypervisor node enabled
in dts. It is necessary action if we want to use driver for Xen
Initial domain shield, that means that the driver isn't tied only
to XenVM board configuration.

Add subdirectory of 'xen' when we have Xen Hypervisor node enabled
in dts. It is necessary action if we want to use arch arm64 core Xen
related code for Xen Initial domain shield, that means that the core
of Xen isn't tied only to XenVM SoC configuration.

Signed-off-by: Mykola Kvach <mykola_kvach@epam.com>