drivers: rcar_mmc: add checking if regulator enabled before disabling it
Add checking if a regulator enabled before disabling it. The SD/MMC
driver doesn't know in which state the regulator has been initialized
or in which state it is now, because other drivers can use it. During
initialization sequence of SD card, first of all, SDHC subsystem
disables power and refcounter inside regulator may become negative
therefore enabling of regulator may not work.

Signed-off-by: Mykola Kvach <mykola_kvach@epam.com>